tdf#140594 Create proper window peer for FORMATTEDFIELD
    commit 676d8f737bbe8458580b94e94e2817e0b547b1b9
    Date:   Wed May 6 14:08:07 2020 -0400

        lok: fix the window type of the formatted field control

had introduced the new 'WindowType::FORMATTEDFIELD' and set
it for the 'FormattedField' control (instead of using
the previous default value set in the 'SpinField' ctor,
which is 'WindowType::SPINFIELD').

So far, this type was not explicitly handled when
creating the XWindow peer.

Handle it just the same as 'WindowType::SPINFIELD'
and return a new 'VCLXNumericField', which e.g.
makes sure that a "proper" accessible is created
and thus alues of a spin field are announced by
the NVDA screen reader (again).

(See also

    commit 9425ad1b6998996cf2b0e14ceca5b76499b9190c
    Date:   Fri May 29 16:53:59 2020 +0200

        tdf#133498 Make native drawing of FormattedField work (again)

which had fixed a similar regression elsewhere.)

Change-Id: Ie8a4d279a12a1b83a9cf97f1304d84214b3a2a2f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/118568
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>