vcl: nerf Window::SetParentToDefaultWindow()
There is a problem with keyboard focus in an Eclipse RCP application on
WNT; it is apparently caused by re-parenting calling
ImplSalReCreateHWND() and that first transfers the focus to the newly
re-created window and when that dies shortly thereafter to the Eclipse
top-level window instead of LO child window.

Re-parenting doesn't work anyway because mpAppWin is always null outside
of unit tests and mpDefaultWin lives outside of time and space and
doesn't know about any window the LO window may be embedded in.

Re-parenting appears to be unnecessary because if a parent dies with
undisposed children that's considered a bug anyway nowadays.

(partially regression from 1fb7d71d8b4f459b15c5f7ab7e538807790346b1
 which removed conditionals that limited the reparenting to floating
 windows)

Change-Id: I7a3997d5e714b40918337207ec1387922f86f8e6
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/118588
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>