sw XHTML import: improve handling of <object> with images
The reqif/xhtml export can map images to a pair of <object> elements:
the outer one contains the real image and the inner one contains a PNG
fallback.  We used to import this back to OLE objects, except for PNG,
which has a single <object> element.

Improve this by allowing the pair of <object> elements as well: if the
outer type is GIF, then map this to an image and ignore the inner PNG.

Change-Id: I5c56720cf6a054208cd0478a54bdac15ffece9aa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/119221
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins