framework: fix multiple-free in VclToolBarManager::Destroy()
Multiple ImplToolItem with mnId 1000 are inserted in the same ToolBox
and the problem is that only the first one gets its mpUserData set,
because setting and getting that is based on the assumption that mnId is
unique per container.

So in Destroy() for every item with mnId 1000 the same AddonsParams
object is deleted.

The problem is that in ToolBarManager::FillToolbar() the local nItemId
is copied, so its value never changes when iterating the loop because
only its copy is incremented.

(regression from 8ab1b18e8c947e8c994a91fa3a10f9dd385b71ff)

Change-Id: Ica47051f9447c2b9d1981f8373d754874f0f31b0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/119228
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>