crashreport writerfilter: findColumn(unsigned long,unsigned long)
Although the crash reports don't show getRow as a separate
item in the list of functions leading to the crash,
it just CAN'T be a problem with mTableDataStack.empty()
or !mTableDataStack.top() because they are unconditionally
used or popped later on in endLevel().

So I resisted the urge to test everything, and just verified
that the row (which exists for m_aCellProperties) actually
exists in the table manager - similar to getGridBefore().

Change-Id: I5a27dc40db18f05ad4656789d3e61ec1ff4de471
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/119420
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>