tools: use upper case GetWidth/GetHeight for Rect for repr. string
When a rectangle is written to a string stream (usually used for
representation string in CPP unit tests) we used getWidth and
getHeight. It turns out that this methods don't return the correct
value for width / height (off by one) in many cases so instead we
need to use the upper-case variants GetWidth/GetHeight.

Change-Id: Id655236155f94e092d9e9c24f9bd6857e9b0a558
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/119489
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>