tdf#143476 improve import of lockedCanvas
A lockedCanvas is used in Word 2007 to include dml-shapes. It is
treated similar to a group, only that its content is not editable.

This patch changes the implementation so, that the lockedCanvas is
imported as group. Prior only content was imported in the case the
content was a single custom-shape or a single group.
That has resulted in these errors:
Image or line in a lockedCanvas were not imported at all.
Only one of several shapes in a lockedCanvas was imported.

I have changed mpShape to mpShapePtr to reflect the data type and be
consistent with ShapeContext.

The patch changes GraphicImport so, that lockedCanvas as special
version of a group is detected and handled. That solves regression
from commit 61a6be3aa3b470e5da3ca7ad4b8d1a6155a6a07f

Change-Id: I57005e64ff073bca6ebb46a404486203fb6802ac
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/119684
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>