tdf#143657 'execute' button doesn't get enabled when contents changed
since...

commit 88e930ee20d806549569fd88a9011ca2d1d214f1
Date:   Tue Jul 6 18:51:38 2021 +0200

    tdf#132740 don't broadcast if modified status has not changed

now we only get a notification on transition from unmodified to
modified. So continue to launch a timer on transition but clear
modification on firing so we will get notified on the next change
and move the modify callback to the timer.

modifications are now deferred until the timer fires, so reduce
the timer to make it smoother

Change-Id: I27980be54de20bb516b1f106c3ed1b303e2a4014
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/119933
Tested-by: Caolán McNamara <caolanm@redhat.com>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>