tdf#143722 DOCX import: map using default TOC heading style
of Writer ("Contents Heading") instead of "TOC Heading".

This essentially reverts the fix to bug #83300.

When that was added the TOC import was broken enough that
it made no visible regression.

In 6.3 the TOC import was fixed up so now it's more annoying.

The unit test of bug #83300 is reused but depends on the fix to
bug #143726.

I checked the example file of bug #83300 and the layout problem
of comment 6 that was fixed there does not come back.

Follow-up to commit 826ce6d804a6c776d37eae21c6dba10d2c223d85
"tf143726 DOCX: export default TOC Header style with correct name".

Change-Id: Iaee73729ea46a0c36d08ccc6fc5fb04fdf592a1d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/120092
Tested-by: László Németh <nemeth@numbertext.org>
Reviewed-by: László Németh <nemeth@numbertext.org>