avoid repeated writer layout calls with tiled rendering
My next commit will disable idle paints for tile rendering, as they
do nothing in that case (or at least that should be the case).
But this code causes an infinite loop with idle layout timer,
because painting resets the SwViewShellImp paint region, which is
not performed if painting is done only later by idle rendering.

Change-Id: Ic401f16254aade02ddb3e4faffd99f0ce29d4df4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/120201
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>