sw HTML export: avoid writing EMF data with JPG extension
Regression from commit 9377eb2bd6a7b8cebcb43b68dfed250d58d1b467 (sw
XHTML import, improved <object> handling for images: support more
formats, 2021-07-20), the problem was that now an EMF image was written
with the JPG extension in the HTML export.

Solve the problem by making the "use emf as-is" check more strict in
XOutBitmap::WriteGraphic(): that will be still true in the XHTML case
when the source image is EMF, but the filter name will be JPG in the
HTML case, so that will avoid the unwanted EMF data in the HTML case.

Change-Id: I7bced482d677cf121be1ed494a3aaeb319e1c5b4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/120209
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins