tdf#143574 sw: textboxes in group shapes - part 1
Introduce SwTextBoxNode class to support grouped
textboxes, fixing the crash when entering a group
shape, trying to add a textbox to one of the shapes.

Test of crash fix: right click on a group shape.
Select the menu item "Enter group". Select one of
the shapes, and right click on it, and choose "Add
Text Box".

Note: textboxes in Writer are linked in SwFrameFormat
class. Each textbox consists of a text frame and a shape.
This object pair makes it possible to have complex
content inside any kind of shape (pictures, tables etc.
Group shapes have only one SwFrameFormat, but can have
many shapes so that means they can only have one textbox.
From now, each shape could have a textbox in a group.
Please note this is only a preparation for that. Filter
implementation and handlers are under development.

Change-Id: Iae35c118f0e67697b289c30d0fad4f5e16501c02
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/120452
Tested-by: László Németh <nemeth@numbertext.org>
Reviewed-by: László Németh <nemeth@numbertext.org>