bibliography, local URL: better starting directory when picking a new file
Improve the case when we have no old file, so we can't take its old
parent directory as a starting point in the file picker, opened by the
browse button.

It turns out both places that open that file picker have access to the
base URL of the current document, so can open the directory of that
instead.

Change-Id: I4db84b6036771f48962d49ae5e4c46a99c306def
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/120455
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins