tdf#143904 fix direct character formatting copied into paragraph level
Currently, Writer copied the direct character formattings, and applied them on paragraph level, when inserting a new row in a table.
This bug is a regression from b3e89c84779d40947d55e91fdd87fddfcf84be1e .

Explanation, why this regression happened:
MSO Word uses a paragraph mark symbol (i.e the control character than can be configured to be shown), which is imported in Writer as a 0-length hint.
To fasten up the import of DOCX tables, the core applies the formattings from this 0-length attribute on paragraph level, directly before inserting a new row.
All in all, to solve this bug, we needed to make sure, that we only copy the formattings, if it is from a 0-length hint.

Change-Id: Iba52e9318cc588feda559339e42897568d009806
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/120624
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>