save LastTipOfTheDayShown when the dialog is shown not in its dtor
Otherwise if a tip dialog is left open then its considered not to be
have be shown and new ones will constantly open on every new
document. Especially problematic in impress if you leave the first
one open and select a template to open from the template dialog then
you get two of them in the same frame.

Change-Id: I94cd34aa031e133d8c229a0de78582fda1dbdf4a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/120674
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>