tdf#142264: make sure to load potentially unloaded objects when saving
Commit 04fba5e3aa2d77b635a4588551bfb85c8cfcfc09 happened to reveal
a pre-existing problem that XLSX export only saved those OLE objects
that were kept loaded in the OLE object cache, subject to thevalue of
org.openoffice.Office.Common/Cache/DrawingEngine/OLE_Objects.

Before that change, the imported charts were marked modified on load,
and that prevented them from unloading in OLEObjCache::UnloadCheckHdl,
because SdrOle2Obj::CanUnloadRunningObj returned false.

After the mentioned change, the charts started to load without the
wrong "modified" state, which allowed them to be properly managed by
the cache, and the export filter implementation error surfaced. It's
likely that commit 945e04eb144632f8690f153a97d7bb5ea2638b77 tried to
workaround the same underlying problem for charts that for some reason
/ at some point in time didn't get marked modified on load, and that
commit converted an error shown in Excel into silently missing charts.

This change makes sure that whenever a reference to chart document
is requested from XclExpChartObj, it is actually loaded and ready
for reading data.

Possibly something could be done on the level of old reference that
becomes non-functional (although valid) as the result of unloading,
so that it would automatically reload on following use. That would
make operating on the references robust. I didn't find an obvious
way to do that.

It is interesting to investigate, it the heizenbug related to images
disappearing from documents, as users keep reporting without robust
reproducers, might possibly be caused by a similar problem.

Change-Id: I45fcdc98254157d805c7519340b5265526f27166
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/120688
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>