vcl: make OutDevState constructor default, no need for destructor
There is no value of having the OutDevState constructor and destructor
being in its own file, so move to header. Also, resetting the variables
in the destructor is rather pointless as they are destroyed in the next
step anyhow. This removes the need to define an empty destructor.

Use C++11 magic to make a default constructor - many thanks to Mike
Kaganski for the tip on this one.

Change-Id: I9561b311d8752fa12802b5ae8e11123aedabe8be
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/115460
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>