tdf#144037 Qt5 just close popup with mouse outside
How hard could it be to close a popup window when losing focus or
clicking outside of it (AKA tdf#143114 problem 3)? At least with
Qt it feels like solving an NP problem...

So don't always force-close the popup window if clicked anywhere
on it. Seems to be consistent with gen and gtk3.

Change-Id: Ia1c231c0b51a55bec2c587feffcf14bb4b4cee55
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/121093
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>