tdf#126126 ODT import: improve import of cross-table commented text range
Regression from commit 06b0fbcb8b2b10fbbbe6a7cb00b2e76ff6e03501
(office:annotation-end import, 2012-07-20), the problem was that
XMLAnnotationImportContext::EndElement() assumed that we can always call
gotoRange() to go from the annotation start and end points, but this is
not true: an annotation may start inside a table and end outside a
table, which is not a valid selection, so gotoRange() fails.

Fix the regression part by just creating a text range for the anchor of
the comment, so the comment is attached to the end of the range, and
this way the rest of the comment & the document can be at least opened.

[ It seems bookmarks behave similarly: they don't block the whole
import, but don't work cross table boundaries, either. ]

Change-Id: I1b5a2e2e7501ce3054379fc79d2045c3439c52e4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/121322
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>