sfx2: allow storeAsURL() on the main thread
This is similar to commit f392f8bd2a25f6ee243812e8a7a2c60ad49304db
(sfx2: allow storeToURL() on the main thread, 2019-02-05), just this
handles the 'export' case, not the 'save as' one.

Note that this already worked in some cases when impl_store() decided to
call storeSelf(), but vcl::solarthread::syncExecute() already avoids
dispatch when called on a non-main thread, so that is not a problem.

Change-Id: Ief09ba6c60923c7f16d86f61625c0faa9a5f829b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/121385
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins