tdf#137363 DOCX filter: no NUMBERING_LEVEL 9(no-level) in styles
The only real (initial) reason for setting PROP_NUMBERING_LEVEL
in the style was to round-trip it so it didn't get lost.
It does kinda give a nice side benefit in UI, where the style
can now set the listLevel at the same time as turning numbering on
(at least on the second try.)

However, 9 ends up meaning LEVEL 10 in LO UI, instead of NO LEVEL.
So eliminate that value from round-trippability,
since LO has no correlating NO LEVEL value possible for that PROP.

The UI is too broken and inconsistent for any attempt
at a unit test.

Change-Id: Ibd9cea2ec10e85651f9f6643ad1b49ebf9babe12
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/121448
Tested-by: Jenkins
Reviewed-by: Justin Luth <justin_luth@sil.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>