tdf#142458 Set correct ScrolledWindow BorderWidth in the ctor
and keep it updated when the relevant properties that affect it change.
This avoids the problem that the borderwidth can change during "Paint"
(which is something we strongly want to avoid) since...

since:

commit 1a07b1ade15cac66043119ecbd6e97673d9edc21
Date:   Thu Nov 19 11:51:13 2020 +0100

    tdf#138010 (IV) VclScrolledWindow: Use actual border width

hopefully after this we can then double-check to see if we can remove
the potential change of BorderWidth during paint.

Change-Id: I2ec317d86687fdb75e6323905f6d1c3b8fc655e1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/121615
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>